Move helpers out of LogicalPlanOptimizer
